--- src/homework/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "ex4",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "nodemonConfig": {
-    "ignore": [
-      "todoDB.json"
-    ]
-  },
-  "dependencies": {
-    "axios": "^0.27.2",
-    "cors": "^2.8.5",
-    "express": "^4.18.1",
-    "fs": "^0.0.1-security",
-    "mysql2": "^2.3.3",
-    "nodemon": "^2.0.16",
-    "sequelize": "^6.21.0"
-  },
-  "devDependencies": {
-    "sequelize-cli": "^6.4.1"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon server.js",
-    "start": "node server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
