--- src/homework/server/db/config/20220622173204-create-item.js
@@ -1,25 +1,0 @@
-'use strict';
-module.exports = {
-  async up(queryInterface, Sequelize) {
-    await queryInterface.createTable('Items', {
-      id: {
-        allowNull: false,
-        autoIncrement: true,
-        primaryKey: true,
-        type: Sequelize.INTEGER
-      },
-      isPokemon: {
-        type: Sequelize.BOOLEAN
-      },
-      item: {
-        type: Sequelize.STRING
-      },
-      isDisplay: {
-        type: Sequelize.BOOLEAN
-      },
-    });
-  },
-  async down(queryInterface, Sequelize) {
-    await queryInterface.dropTable('Items');
-  }
-};
